+++ tvsd/cli.py
@@ -3,9 +3,8 @@ import os
 from pathlib import Path
 import shutil
 from typing import Optional
-from tvsd._variables import BASE_PATH, TEMP_BASE_PATH, SERIES_DIR
+from tvsd._variables import state_base_path, state_temp_base_path, state_series_dir
 from tvsd import state
-
 
 import typer
 from rich import print as rprint
@@ -104,13 +103,22 @@ def main(
 
 
 @app.command()
-def search(query: str):
+def search(
+    query: str,
+    specials_only: Optional[bool] = typer.Option(
+        False,
+        "--specials",
+        "-s",
+        help="Download Specials Only",
+    ),
+):
     """Search for media and download
 
     Args:
         query (str): query string
     """
-    search_media_and_download(query=query)
+    validate_config_file()
+    search_media_and_download(query=query, specials_only=specials_only)
 
 
 @app.command()
@@ -119,10 +127,10 @@ def clean_temp():
     validate_config_file()
 
     try:
-        dir_content = os.listdir(TEMP_BASE_PATH)
+        dir_content = os.listdir(state_temp_base_path())
         if len(dir_content) == 0:
             raise (FileNotFoundError)
-        rprint(f"{TEMP_BASE_PATH} contents: ")
+        rprint(f"{state_temp_base_path()} contents: ")
         for item in dir_content:
             rprint(f"  {item}")
 
@@ -132,11 +140,11 @@ def clean_temp():
             default="n",
         )
         if confirm.capitalize() == "Y":
-            shutil.rmtree(TEMP_BASE_PATH, ignore_errors=True)
-            os.mkdir(TEMP_BASE_PATH)
+            shutil.rmtree(state_temp_base_path(), ignore_errors=True)
+            os.mkdir(state_temp_base_path())
             logging.info("All files deleted")
     except FileNotFoundError:
-        logging.info(f"Temp directory {TEMP_BASE_PATH} does not exist")
+        logging.info(f"Temp directory {state_temp_base_path()} does not exist")
 
 
 @app.command()
@@ -161,7 +169,9 @@ def remove_show():
         if choice < num_rows:
             if typer.confirm(f"Will remove {shows[choice]}. Are you sure?", abort=True):
                 typer.echo("Removing show: " + shows[choice])
-                shutil.rmtree(os.path.join(BASE_PATH, SERIES_DIR, shows[choice]))
+                shutil.rmtree(
+                    os.path.join(state_base_path(), state_series_dir(), shows[choice])
+                )
                 typer.echo("Show removed")
             break
 
@@ -192,7 +202,7 @@ def clean_base(
         help="Remove directories without videos",
     ),
     target: Optional[str] = typer.Option(
-        os.path.join(BASE_PATH, SERIES_DIR),
+        os.path.join(state_base_path(), state_series_dir()),
         help="Target directory",
     ),
     _no_confirm: Optional[bool] = typer.Option(